+++ packages/webpack-ui/src/pages/default.ts
@@ -239,6 +239,7 @@ export default function (discovery: StatoscopeWidget): void {
                             compilation.modules.[not shouldHideModule()].[
                               name~=#.filter or modules and modules.[name~=#.filter]
                             ]
+                            .sort(getModuleSize($hash).size desc)
                             `,
                             content: {
                               ...modulesTree(),