+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next
+## 5.24.0 (17 February 2022)
 
 ### Features